# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {},
		"securitySchemes": {}
	},
	"info": {
		"title": "top-car-api",
		"version": "1.0.0",
		"description": "api top-car",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Lucas Purkota"
		}
	},
	"paths": {
		"/api/usuario/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"senha": {
										"type": "string"
									},
									"cidade": {
										"type": "string"
									},
									"dataNascimento": {
										"type": "string",
										"format": "date-time"
									},
									"celular": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"nome": {
										"type": "string"
									},
									"cpf": {
										"type": "string"
									}
								},
								"required": [
									"senha",
									"cidade",
									"dataNascimento",
									"celular",
									"email",
									"nome",
									"cpf"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/usuario/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"senha": {
										"type": "string"
									},
									"cidade": {
										"type": "string"
									},
									"dataNascimento": {
										"type": "string",
										"format": "date-time"
									},
									"celular": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"nome": {
										"type": "string"
									},
									"cpf": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"senha",
									"cidade",
									"dataNascimento",
									"celular",
									"email",
									"nome",
									"cpf",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/usuario/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/usuario/getAll": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/usuario/getEmail": {
			"post": {
				"operationId": "GetEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"senha": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"senha",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/venda/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"vendedor": {
										"type": "string"
									},
									"valor": {
										"type": "number",
										"format": "double"
									},
									"detalhes": {
										"type": "string"
									},
									"combustivel": {
										"type": "string"
									},
									"km": {
										"type": "number",
										"format": "double"
									},
									"ano": {
										"type": "string"
									},
									"modelo": {
										"type": "string"
									},
									"marca": {
										"type": "string"
									}
								},
								"required": [
									"vendedor",
									"valor",
									"detalhes",
									"combustivel",
									"km",
									"ano",
									"modelo",
									"marca"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/venda/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"vendedor": {
										"type": "string"
									},
									"valor": {
										"type": "number",
										"format": "double"
									},
									"detalhes": {
										"type": "string"
									},
									"combustivel": {
										"type": "string"
									},
									"km": {
										"type": "number",
										"format": "double"
									},
									"ano": {
										"type": "string"
									},
									"modelo": {
										"type": "string"
									},
									"marca": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"vendedor",
									"valor",
									"detalhes",
									"combustivel",
									"km",
									"ano",
									"modelo",
									"marca",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/venda/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/venda/getAll": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/venda/getSaleUser": {
			"post": {
				"operationId": "GetSaleUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"vendedor": {
										"type": "string"
									}
								},
								"required": [
									"vendedor"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}